MYM-3 Create ports

+++ src/main/java/com/prokopchuk/mymdb/domain/UserRating.java
@@ -1,4 +1,4 @@
 package com.prokopchuk.mymdb.domain;
 
-public record UserRating(User user, Film film, double rating) {
+public record UserRating(User user, Film film, int rating) {
 }
